Removed external fonts, and changed to with directly with emoji characters
Removed external fonts and any character blocks that needed them.
Modified script to deal directly with emoji characters, by intercepting the post event, and converting non-BMP characters to html entities at that point. This required modifying the preview routine to no remove non-BMP characters - this is only done if this script is enabled.
Added a setting to enable the custom emoji keyboard. A message for how to activate the system emoji keyboard is now always shown.