Added my greasemonkey chatty news script to ChromeShack
Add Shack news recent headlines to top right of chatty.
Optionally adds article description the Shacknews chatty article posts, and can highlight these posts.
Optionally adds a 'wallpaper' background based off of a dimmed version of the main image from the most recent article. This image can optionally be made full brightness at the top and bottom of the chatty so it can be seen clearly. This adds an intereating game related image to the chatty background, and is useful as if there is a different image when you refresh the chatty, you know a new article has been posted.